[app-names-refactor] Model naming changes
