+++ Locus/settings.py
@@ -38,8 +38,8 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'search.apps.SearchConfig',
-    'tweetCrawler.apps.TweetCrawlerConfig',
-    'googleCrawlerOfficial.apps.GoogleCrawlerOfficialConfig',
+    'twitter.apps.TweetCrawlerConfig',
+    'searchEngine.apps.GoogleCrawlerOfficialConfig',
     'database.apps.DatabaseConfig',
     'channels',
     'rest_framework',
