+++ search/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 
 from search.models import SearchParameters
-from tweetCrawler.models import TwitterUser
+from twitter.models import TwitterUser
 
 admin.site.register(SearchParameters)
 admin.site.register(TwitterUser)
