+++ searchEngine/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from googleCrawlerOfficial.models import InternetResult, Domain
+from searchEngine.models import InternetResult, Domain
 
 admin.site.register(InternetResult)
 admin.site.register(Domain)
