+++ searchEngine/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class GoogleCrawlerOfficialConfig(AppConfig):
-    name = 'googleCrawlerOfficial'
+    name = 'searchEngine'
