+++ searchEngine/patterns.py
