+++ twitter/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class TweetCrawlerConfig(AppConfig):
-    name = 'tweetCrawler'
+    name = 'twitter'
